+++ composeApp/src/commonMain/kotlin/io/github/sophon/fightingnerd/feat/quiz/ui/overview/QuizOverviewState.kt
@@ -1,7 +1,9 @@
 package io.github.sophon.fightingnerd.feat.quiz.ui.overview
 
 import io.github.sophon.fightingnerd.feat.quiz.model.QuizGameWidget
+import kotlinx.collections.immutable.ImmutableList
+import kotlinx.collections.immutable.persistentListOf
 
 internal data class QuizOverviewState(
-    val quizGameWidgetList: List<QuizGameWidget> = emptyList(),
+    val quizGameWidgetList: ImmutableList<QuizGameWidget> = persistentListOf(),
 )