+++ composeApp/src/commonMain/kotlin/io/github/sophon/fightingnerd/feat/quiz/ui/overview/QuizOverviewVM.kt
@@ -3,6 +3,7 @@ package io.github.sophon.fightingnerd.feat.quiz.ui.overview
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import io.github.sophon.fightingnerd.feat.quiz.usecase.SubscribeGameWidgetsUseCase
+import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
@@ -23,7 +24,7 @@ internal class QuizOverviewVM(
     private fun loadWidgets() {
         viewModelScope.launch {
             subscribeGameWidgetsUseCase.invoke().collect { gameWidgetList ->
-                _state.update { it.copy(quizGameWidgetList = gameWidgetList) }
+                _state.update { it.copy(quizGameWidgetList = gameWidgetList.toImmutableList()) }
             }
         }
     }