+++ composeApp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # APP CHANGELOG
 
+## [v2.2.0] - 2026-08-22
+- `Home` - don't collapse game widget after navigating back 
+- `Move`
+  - image carousel
+  - image, video and note indicators for expansion
+- `Quiz` - text indicator if media absent
+- further stability improvements
+
 ## [v2.1.0] - 2026-08-20
 - `Wavu` - added Bob
 - `DustLoop` - enable MTFS